Python/Factory/requirements.txt
django
djangorestframework==3.16.1
django-cors-headers==4.9.0
